[h2v] Added a new job property to override store-level rewind time for the new version
Both ETL+VPJ and KafkaInputFormat support in VPJ has the requirement to adjust the rewind
time for the re-pushed version because of the following reasons:
This property comes to play when the default rewind time configured in store-level is too short or too long.
1. If the default rewind time config is too short (for example 0 or several mins), it could cause data gap with
re-push since the push job itself could take several hours, and we would like to make sure the re-pushed version
will contain the same dataset as the source version.
2. If the default rewind time config is too long (such as 28 days), it will be a big waste to rewind so much time
since the time gap between the source version and the re-push version should be comparable to the re-push time
if the whole ingestion pipeline is not lagging.

There are some challenges to automatically detect the right rewind time for re-push because of the following reasons:
1. For Venice Aggregate use case, some colo could be lagging behind other prod colos, so if the re-push source is
from a fast colo, too short rewind time could cause a data gap in the slower colos. Ideally, it is good to use
the slowest colo as the re-push source.
2. For Venice non-Aggregate use case, the ingestion pipeline will include the following several phases:
2.1 Customer's Kafka aggregation and mirroring pipeline to replicate the same data to all prod colos.
2.2 Venice Ingestion pipeline to consume the local real-time topic.
We have visibility to 2.2, but not 2.1, so we may need to work with customer to understand how 2.1 can be measured
or use a long enough rewind time to mitigate all the potential issues.

New job property: rewind.time.in.seconds.override
By default, it is -1, and the new version will pick up the store-level rewind time config.

RB=2557752
BUG=VENG-7471
R=jlliu,nithakka,ssen,yurzhou,lcli,xnma,smaji,fvillega,xhao,shuhliu,xyin,sdwu,zpolicze,anaberez,mnhuang
A=zpolicze