[da-vinci] Split ingestion backend interface for Da Vinci and Venice Server
This RB splits the original IngestionBackend interface into two interfaces: DaVinciIngestionBackend and VeniceIngestionBackend for different use cases, as some of the APIs are designed for only one of the product and splitting them make it more clear for future development.
These two interfaces will extends a shared interface: IngetsionBackendBase, which contains the shared APIs for both products.
DefaultIngestionBackend and IsolatedIngestionBackend implemenets both interfaces so they can be used in both products.

RB=2560166
BUG=VENG-7528
G=voldemort-reviewers
R=nithakka,xyin,yurzhou,goliu,zpolicze,shuhliu,mnhuang,lcli,sdwu,ssen,smaji,xnma,anaberez,fvillega,xhao
A=goliu