[server] Avoid blindly creating VeniceWriter
A VeniceWriter(VW) is needed in Leader/Follower ingestion task in 2 cases:
1. Leader needs to replicate data from real-time topics to version topics (hybrid store)
2. Native replication is enabled and leader is consuming data remotely.

However, currently there is a bug in Leader->Standby transition where
a VW is always created in order to send out End_of_Segment (EOS)
message for a partition. As a result, even though the store is not hybrid
and native replication is not enabled, an extra VW is created as long as
there is a Leader->Follower transition (whenever there is a leadership handover).

This RB fixed the bug by not creating a VW for the sake of sending EOS, and only
send EOS if a VW already exists.

RB=2560393
G=voldemort-reviewers
R=xyin,sdwu,nithakka,smaji,yurzhou,fvillega,ssen,xhao,shuhliu,xnma,lcli,zpolicze,jlliu,goliu,anaberez
A=fvillega