[server] Adding cluster level aggregrated lag metrics
This is introducing following additional lag metrics. This will be useful
in monitoring the system better.
1. agg_future_leader_offset_lag: This will report sum of batch_leader_offset_lag
   for future version for all ongoing push jobs.
2. agg_future_follower_offset_lag: This will report sum of batch_follower_offset_lag
   for future version for all ongoing push jobs.
3. agg_future_batch_replication_lag: This will report sum of batch_replication_lag
   for future version for all ongoing push jobs.
4. batch_replication_lag: This will report difference between offset in
   remote VT and local VT.

For the above aggegated lag metrics(1,2,3) we are not looping over all the ingestion
tasks again. Since each of the ingestions task's individual lag metrics will be
calculated always it can make use of that to aggregrate it and then reset the counter.
Ingrpahs peridoically (every 1 min) poll all the gauge metrics.

RB=2558282
BUG=VENG-7412,VENG-7413,VENG-7411
G=voldemort-reviewers
R=sdwu,xnma,shuhliu,lcli,smaji,xhao,anaberez,nithakka,zpolicze,goliu,mnhuang,jlliu,fvillega,xyin,yurzhou
A=mnhuang