[controller] Standby->Leader controller performing Leader operations and failing leadership check
Problem:
During the Standby -> Leader transition the controller is asked to perform Leader operation as part of
the transition (if some conditions are met). In this specific failing scenario it was
VeniceHelixResources.refresh -> PushMonitorDelegator.loadAllPushes -> AbstractPushMonitor.loadAllPushes
-> AbstractPushMonitor.handleOfflinePushUpdate -> AbstractPushMonitor.handleCompletedPush
-> AbstractPushMonitor.updateStoreVersionStatus.
This call trace is triggered when the push monitor realizes that some pushes can be completed during
loadAllPushes. AbstractPushMonitor.updateStoreVersionStatus will make updates to the store despite
it's not the Leader of the cluster yet. The update is successful because (1) we explicitly refreshed
metadataRepository in VeniceHelixResources.refresh first and (2) there is no leadership check within
the metadataRepository. However, as part of the recent metadata store writer addition we attempt to
retrieve the metadata schema id via VeniceHelixAdmin which enforces leadership and during the trnasition
the controller is not the Leader yet.

Fix:
The provided fix is a simple patch/hack that bypass the leadership check by using the schema repository
directly to get schema id. This is acceptable because MetadataStoreWriter is going to be replaced by
MetaStoreWriter and the way we manage schemas in meta system store does not require us to retrieve schema
id dynamically. However this fix doesn't protect the Standby -> Leader transition from other unknown
call trace that implicitly assumes the controller is Leader already.

RB=2561394
BUG=VENG-7512
G=voldemort-reviewers
R=shuhliu,zpolicze,ssen,fvillega,mnhuang,lcli,xnma,smaji,yurzhou,xhao,sdwu,nithakka,jlliu,anaberez,goliu
A=goliu