[da-vinci] Return Optional.empty() instead of throwing exception at StorageMetadataService.getStoreVersionState() when storage engine is removed
When a topic is removed from Da Vinci, the storage engine will be removed. However, the version level offsetLag metrics is not aware of this and still ask for StoreVersionState from metadata partition and throw Exception for a short period of time and cause confusion.
This RB changes the behavior in `StorageMetadataService.getStoreVersionState()` to catch exception and return Optional.empty() when storage engine is not found.

RB=2568116
BUG=VENG-7543
G=voldemort-reviewers
R=xyin,shuhliu,anaberez,zpolicze,lcli,xhao,sdwu,yurzhou,mnhuang,fvillega,smaji,ssen,goliu,xnma,nithakka
A=anaberez