[test][controller] Fix issue in measuring last push duration metric on controller leadership switchover.
If a push job completes very fast and the job duration is calculated to be
0 seconds then on controller leadership switchover the new leader were not recording the
"successful_push_duration_sec_gauge" when it loads all the push
job details again from ZK. This was causing the integration test
testControllerRestartFetchesLastSuccessfulPushDuration to fail randomly
and be flaky.
Though this issue is practically not possible it's good to fix this by
just recording all non-negative values for the last successful push duration on
leadrship switchover.
This will also make the test reliable.

RB=2574166
BUG=VENG-7400
G=voldemort-reviewers
R=xnma,xyin,sdwu,zpolicze,yurzhou,lcli,fvillega,jlliu,xhao,nithakka,smaji,shuhliu,mnhuang,anaberez,goliu
A=nithakka