[da-vinci][thin-client][common][router] Add new store state router endpoint for system store discovery
The auto system store current version discovery feature is 80% complete after this change. To reduce RB side
and make it easier to review I've split the other 20% to another RB. The remaining work is the ability to
detect current version change after the DaVinci client is up and running.

Changes:
1. Added a new router metadata endpoint to retrieve Store object(state) from zk repositories. The endpoint is
capable of retrieving states for both regular and system stores.

2. Created a path to serialize SystemStore by creating a wrapper class SerializableSystemStore.

3. Refactored SchemaReader a bit in order to reuse part of it in the new StoreStateReader. The client which
calls the router StoreState endpoint.

RB=2578792
BUG=VENG-7344
G=voldemort-reviewers
R=shuhliu,jlliu,goliu,smaji,xnma,nithakka,xhao,mnhuang,zpolicze,fvillega,lcli,sdwu,yurzhou,anaberez,ssen
A=goliu