[h2v][router][da-vinci][server][schema-common][thin-client][fast-client][test] Use CompressorFactory and ChunkingAdapter instances instead of static methods to use separate objects per component
The issue with the CompressorFactory is that as a Factory method, it creates only one instance of the dictionaries.
Although this is desirable, problems arise in tests where all components share the same process space and hence,
run into race conditions if one component closes the compressor while another is using it.

For example, it is desirable for mapper to close the compressor after the map tasks are done. But in tests, this
ends up closing the compressor instance that the routers are using.

This RB creates a single instance of CompressorFactory per component.

RB=2579892
BUG=VENG-7592
G=voldemort-reviewers
R=ssen,xnma,sdwu,xyin,fvillega,smaji,shuhliu,zpolicze,lcli,goliu,jlliu,mnhuang,xhao,yurzhou,anaberez
A=yurzhou