[controller] Skip L/F state model check for the NR rollout admin command
Currently L/F state model is not enabled in the parent controller during
production roll-out; however, a safeguard was added into the native
replication (NR) rollout command, to check the prerequisite L/F state model;
as a result, this new command doesn't work in the real prod environment.

Similar issue happened to the previous cluster level config for NR rollout,
which was fixed by introducing a new controller config:
lf.model.dependency.check.disabled
If this config is on, we don't check the prerequisite L/F state model; this
RB fixs the admin command with the same strategy; if the config is on, skip
the L/F state model check.

RB=2591534
BUG=VENG-7611
G=voldemort-reviewers
R=jlliu,shuhliu,goliu,xhao,sdwu,smaji,yurzhou,xnma,adietz,nithakka,xyin,zpolicze,fvillega,anaberez,ssen,lcli
A=ssen