[da-vinci] Transfer LeaderFolowerState from child proces to main process in ingestion isolation
This RB adds support for ingestion isolation to pass LeaderState from child process to main process when ingestion completed in child process.
This effor is needed for server integration as server needs to support L/F ingestion.
For Da Vinci it is not needed, as Da Vinci clients always run in follower mode.

RB=2594014
BUG=VENG-7654
G=voldemort-reviewers
R=sdwu,shuhliu,anaberez,xhao,xyin,zpolicze,yurzhou,ssen,xnma,nithakka,fvillega,goliu,lcli,adietz,smaji,mnhuang
A=goliu