[common][server] Simplify offset measurement api in TopicManager.
Problem #1:
Currently to get the latest and earliest offset of a topic partition we use 4 different KafkaConsumer API in following order
assign(topicpartition1)
seektoEnd(topicPartition1)
position(topicPartition1)
assign(empty)

This is a bit inefficent as we can simply use a single API like endOffsets/beginningOffsets api to get the same info.

Problem #2:
We also saw an issue with thread synchronisation when getLatestOffset throws TimeoutException for topicpartition1 and then it throws this error
"No current assignment for partition for topicpartition1" even when next time it's called with a different topicpartition2. We do not have a proper RCA for this issue yet.

This change will solve the #1 and may help with #1.

RB=2594758
BUG=VENG-7618
G=voldemort-reviewers
R=shuhliu,goliu,xyin,mnhuang,lcli,nithakka,fvillega,smaji,xnma,zpolicze,adietz,anaberez,yurzhou,jlliu,xhao,sdwu
A=mnhuang,lcli