[router] Prevent post dictionary download actions from running in the calling thread
If the result of the CompletableFuture is available by the time CompletableFuture.handle is invoked, handle will run the provided
function synchronously. This leads to the wait in post dictionary download actions to block the execution of the calling thread
(the very condition it is waiting for). Due to this, the while loop is effectively an infinite loop. With handleAsync, we force
the execution of post dictionary download actions in a separate thread.

RB=2595689
BUG=VENG-7504
G=voldemort-reviewers
R=shuhliu,smaji,xhao,anaberez,adietz,lcli,goliu,yurzhou,fvillega,xyin,jlliu,mnhuang,xnma,zpolicze,ssen,sdwu
A=ssen