[da-vinci] Fix subPartition report condition checking issue in PartitionConsumptionState
This RB adds an in-memory hash set which keeps track of all previous status this sub-partition has reported. It is the in-memory cache of the previousStatuses field in PartitionState inside link OffsetRecord.
The reason to have this HashSet is to maintain correctness and efficiency for the sub-partition status report condition checking. The previousStatus is a generated CharSequence to CharSequence map and we need to iterate over all the records in it to in order to compare with incoming status string.
Without explicit locking, this iteration might throw `java.util.ConcurrentModificationException` in multi-thread environments.

RB=2597157
G=voldemort-reviewers
R=nithakka,ssen,smaji,xhao,lcli,goliu,xnma,yurzhou,adietz,sdwu,shuhliu,mnhuang,zpolicze,fvillega,anaberez,xyin
A=xnma