[server] Making hybrid state transition faster after restarting storage nodes.
Currently, when a Venice storage node starts, it will measure the partition replica offset lag and compare it with the pre-set threshold to decide if it is in BOOTSTRAP or ONLINE. However, during steady state for real-time push, if the offset lag > the threshold, the Venice storage node won't convert the status from ONLINE to BOOTSTRAP. This inconsistent transition from BOOTSTRAP to ONLINE in start and steady state will delay the online state switch.

To avoid the inconsistent transition from BOOTSTRAP to ONLINE in start and steady state, we intent to apply the same ready-to-serve criteria for offset lag. Essentially we plan to store the previous steady-state lag info into metadata store exactly before the storage node shuts down. Then after restart, we are able to retrieve the previous steady-state status and decide to turn the partition replica to be ONLINE right away.

Add a config: OFFSET_LAG_DELTA_RELAX_FACTOR_FOR_FAST_ONLINE_IN_RESTART to enable this feature (>0) and setting the acceptable delta of offset lag growth.

Deployment order: since we upgrade the PartitionState in this change, controllers need to be deployed before servers to support forward compatibility feature for this protocol.

RB=2553713
BUG=VENG-7510
G=voldemort-reviewers
R=shuhliu,fvillega,jlliu,xyin,ssen,zpolicze,sdwu,anaberez,lcli,xnma,mnhuang,goliu,yurzhou,smaji,nithakka
A=mnhuang