[controller] Add initiatingCluster parameter to Admin updateClusterDiscovery API
1. updateClusterDiscovery currently always try to get the StoreConfigAccessor through the old cluster name parameter's
VeniceHelixResource. However, in situations such as abort migration the initiating controller is not the leader of the
store's current cluster. e.g. store A is already in destination cluster but abort migration is initiated and the
updateClusterDiscovery will be called in source cluster's leader controller and if the controller is not the leader of
source and destination clusters it will throw exceptions.

The new initatingCluster parameter will be used to obtain the StoreConfigAccessor from the correct VeniceHelixResource.

RB=2598709
BUG=VENG-7641
G=voldemort-reviewers
R=shuhliu
A=shuhliu