[da-vinci][server] Create amplification factor adaptor to clean up logics in Server and Da Vinci implementations
Amplification factor logics scatteres in many places and this will create difficulties in future iterations and integration with other logics.
This RB is the first part of the effort to user adaptor to centralize amplification logics in ingestion path.
After this clean up, Server and Da Vinci could hide amplification logics as much as possible, IngestionBackend will also work on user partition level.
Note that this logic clean up is not complted. For read path logic, the clean up will happen in the future RB.
This will also help ingesiton isolation and server integration cleaner.

RB=2602030
BUG=VENG-7665
G=voldemort-reviewers
R=smaji,nithakka,ssen,xnma,xyin,goliu,sdwu,mnhuang,adietz,xhao,zpolicze,shuhliu,yurzhou,fvillega,lcli,anaberez
A=goliu