[test] Fixed flaky tests in TestHybrid
The integration test case TestHybrid uses a shared Venice cluster
across all individual test cases, and the shared cluster should be
closed in the end of the whole class. However, there is a test
case the would close the cluster; the rest of the tests that run
after it would not be able to past, usually fail with unable to
find master controller.

This RB removes the bug and closes shared cluster at the end of class.

RB=2607278
BUG=VENG-7648,VENG-7647
G=voldemort-reviewers
R=nithakka,xnma,adietz,goliu,smaji,anaberez,shuhliu,zpolicze,ssen,xhao,yurzhou,fvillega,sdwu,xyin,jlliu,lcli
A=xyin,jlliu