[da-vinci][test-common] Fix flaky test DaVinciClientTest#testUnstableIngestionIsolation
When main process restarts the isolated ingetsion process, we should be confident to set the `latestHeartbeatTimestamp` as current timestamp, as restart of the forked process is already completed. Otherwise, if we kill the forked process again in a short time (less time one heartbeat check period), the heartbeat timeout validation logic would not work as the `latestHeartbeatTimestamp` is always -1.

RB=2615072
BUG=VENG-7545
G=voldemort-reviewers
R=anaberez,yurzhou,xhao,shuhliu,nithakka,fvillega,xyin,zpolicze,sdwu,xnma,mnhuang,goliu,ssen,lcli,adietz,smaji
A=xnma