[da-vinci][server][test-common] Make sure metadata partitions are always opened in isolated process when ingestion isolation is on
For now, when live-update-suppression is on, it restores metadata partition on main process instead of forked process. This will cause failure if other partitions are subscribed in the future, they won't be able to find is metadata partition in forked process.
This RB fixes the issue by blocking metadata restore in Da Vinci (Server blocks metadata/data partitions already since we don't need bootstrap there).

RB=2626763
BUG=VENG-7749
G=voldemort-reviewers
R=yurzhou,xyin,smaji,xhao,nithakka,anaberez,adietz,ssen,mnhuang,fvillega,shuhliu,sdwu,xnma,zpolicze,goliu,lcli
A=mnhuang