[server][da-vinci] Update total versioned stat when version list updates
If the stat type in versioned stats is a counter which always increase,
the total stats will keep increasing until the host is bounced.

This RB fixed the issue by updating total stats when a version swap takes
place. The update function will aggregate the value across the new version
list, and update the total stats.

RB=2635813
BUG=VENG-7732
G=voldemort-reviewers
R=jlliu,fvillega,goliu,shuhliu,xhao,nithakka,adietz,anaberez,ssen,zpolicze,sdwu,xnma,smaji,yurzhou,xyin,lcli
A=xyin