[controller] Fixed leaked Brooklin streams
When terminating Brooklin streams, we checked whether the store has
Leader/Follower state model enabled; if so, skip the deletion of
Brooklin streams. However, during L/F migration, the store level
config is definitely on, which caused the leak of Brooklin streams.

This RB fixed the leak by checking the version level configs when
terminating Brooklin streams; if the version is on Online/Offline
state model, terminate the Brooklin stream when the corresponding
Helix resource is retired.

RB=2635834
BUG=VENG-7327
G=voldemort-reviewers
R=fvillega,anaberez,sdwu,zpolicze,lcli,xyin,yurzhou,adietz,ssen,jlliu,shuhliu,xnma,goliu,xhao,nithakka,smaji
A=anaberez,shuhliu