[davinci][controller] Add checkSchema to the partitioner interface
checkSchema is now a function called in the constructor of the VenicePartitioner.  It can be overridden to throw an exception should a provided schema not match.  Today, when a prtitioner is added to a store, the parent controller attempts to instantiate the the partitioner and can throw if something goes wrong.

RB=2636832
BUG=VENG-7714
G=voldemort-reviewers
R=goliu,xhao,smaji,lcli,sdwu,xnma,xyin,mnhuang,yurzhou,anaberez,adietz,nithakka,jlliu,ssen,shuhliu,fvillega
A=anaberez