[davinci][server] Skip writing to the meta system store when the corresponding Venice store is deleted
Previously, the meta store writer won't be removed even the corresponding
Venice store is deleted, and when stopping the Venice Server, the
VeniceWriter for the deleted meta store will still try to close it
by sending out some Control Messages to the deleted RT topic, which
could hang forever with the current config.

This code change is trying to get rid of meta system store writer
when the store is deleted as early as possible, and also adding
a topic existence check before closing VeniceWriter to make it
more flexible.
But there is a race condition there between the topic existence check
and sending message to the topic, but the chance to hit this race condition
should be very low.
If this doesn't work well, we may need to construct a special
VeniceWriter for meta system store writer to avoid using inifinite retry
and retry inside the application code.

RB=2641995
BUG=VENG-7324
R=shuhliu,ssen,fvillega,xnma,jlliu,smaji,xyin,xhao,yurzhou,anaberez,nithakka,mnhuang,zpolicze,adietz,sdwu,lcli
A=mnhuang