[da-vinci][server] Addtional fix for IsolatedIngestionBackend for leader promotion/demotion
Previous bug fix is not correct: it messed up the true/false again when checking if partition is unsubscribed.
This RB fixes it by inverting the method: we now check if partition is being subscribed. The logic is much easier to understand and verify.

With this fix, TestHybrid.testLeaderCanRelease latch will be much less flaky(another flakiness will be address in a second RB), as the version will always be online.

RB=2651146
G=voldemort-reviewers
R=yurzhou,smaji,nithakka,xhao,goliu,ssen,xnma,shuhliu,fvillega,mnhuang,sdwu,zpolicze,adietz,anaberez,lcli,xyin
A=xnma