[test] Fix flaky test TestPushJobWithNativeReplicationSharedProducer
Since this test is running 3 push jobs in parallel sometimes it takes
a little longer (more than 30 seconds) to finish all the jobs. Before
checking if the new version has come online let's wait for the push job
threads to finish. This will make sure that at least the push jobs have
finished before we move on to the verification stage.

RB=2641212
BUG=VENG-7785
G=voldemort-reviewers
R=xhao,zpolicze,adietz,nithakka,goliu,sdwu,lcli,yurzhou,jlliu,fvillega,xnma,shuhliu,smaji,xyin,anaberez,mnhuang
A=mnhuang