[server] Fix mid segment check sum type mismatch with state
In short summary, we found when we meet a mid segment without SOS, the check sum type and
check sum state doesn't match. e.g. check sum type is set to MD5 but check sum state is set to
empty array. This was caused by in codes we updated check sum type only when when initializing a new
producer partition state. So when we update check sum state to empty array, we forgot to update check sum type.
This caused array index out of bound issue.

RB=2649330
BUG=VENG-7794
G=voldemort-reviewers
R=xnma,xyin,mnhuang
A=xyin