[thin-client][fast-client][davinci][venice-server][h2v] Avoid unnecessary fast-avro verification
The fast-avro verification is slow and it is a blocking operation,
and this code change will try to reduce duplicate verification
when multiple store clients are being used in the same process.

This code change also introduced a concurrent test, which reveals
a bug/behavior change of JDK11 since the fast-class compilation
in the thread pool will fail consistently, and I have reported this
issue to the JAVA team, and hope we could get some update soon.

This code change also fixed some long-tail retry related test failure.

RB=2657683
BUG=VENG-7806
R=lcli,xhao,ssen,nithakka,jlliu,anaberez,fvillega,xyin,shuhliu,sdwu,zpolicze,mnhuang,adietz,smaji,yurzhou,xnma
A=smaji