[controller] Fix batch update admin command processing.
The new NR config command CONFIGURE_NATIVE_REPLICATION_FOR_CLUSTER is
a batch update admin command which queues up a admin task for each
store by looping through the storeAdminOperationsMapWithOffset.
But this map is not gurranted to be updated with all the stores those are currently present in
the cluster. This map is updated on demand basis only when there is admin task
queued for a store. Also the map is never deleted when a store is deleted.
Fixing it by looping through the stores present in the store repository instead.

Also there was another issue with skip-admin command. During batch update commmand processing
if multiple stores throws exception then we have to do skip-admin as many times. This is because
it resets the offsetToSkip to sentinel value after every store. Fixing this as well.

RB=2654989
BUG=VENG-7754,VENG-7837
G=voldemort-reviewers
R=fvillega,anaberez,nithakka,sdwu,xyin,adietz,yurzhou,goliu,shuhliu,mnhuang,lcli,xhao,xnma,jlliu,zpolicze,smaji
A=mnhuang