[Controller] Fixed the stores with bad replication factor config
Several metadata repository factoring caused that some stores are
using '0' as the replication factor, which will fail the new resource
creation, and right now, the new resource creation is only looking
at the store-level replication factor config.

There is a cluster-level config as well, which is being used
during the new store creation.
We have been having some discussion about whether we should use
the cluster-level config as the defaule value if the store-level
config is invalid, but we didn't get a clear conclusion on this.
There is one concern with this approach is that not every component
has the access to the default cluster config, which is managed by
Controller and the backfill/override needs to be used for each access,
which is a little tricky.

So far, I think repairing the store replication factor is the right
balance considering the current situation.

This code change also updated one flag of System store type:
BATCH_JOB_HEARTBEAT_STORE since it is not using zk shared store
structure, so no need to adopt the arch for the zk shared
system stores.

RB=2668480
BUG=VENG-7706,VENG-7843
R=jlliu,mnhuang,zpolicze,shuhliu,adietz,yurzhou,xnma,xhao,sdwu,nithakka,smaji,ssen,fvillega,anaberez,xyin,lcli
A=lcli