[server] Catch newly added fatal exceptions in DIV
Recently we found there is a potential exception thrown in CheckSum. This exception
was not caught and caused ingestion to fail fast. We should catch this fatal exception
and let caller StoreIngestionTask decide what to do next.

After this change, all exceptions in DIV should be able to get caught and processed in StoreIngestionTask #validateMessage.

RB=2672496
BUG=VENG-7838,VENG-7817
G=voldemort-reviewers
R=mnhuang,jlliu
A=mnhuang