[da-vinci-client][server] Got rid of `synchronized` from some functions in AbstractStorageEngine
Two functions in AbstractStorageEngine are in the hot path of both read and write:
getStoreVersionState: critical read path
getPartitionOrThrow: crtical read path and write path

Previously, they are synchronized with other functions in the same class, and they
will be delayed by other heavy operations, such as `sync` and `adjustStoragePartition`
and more others, which will cause latency spikes.

To avoid the following race condition for write-compute enabled stores, this code
change added a read-write lock:
This lock is used to guard the re-opening logic in {@link #adjustStoragePartition} since
{@link #getPartitionOrThrow} is not synchronized and it could be invoked during the execution
of {@link #adjustStoragePartition}.

The reason to introduce this rw lock is that for read-compute enabled stores, the lookup could happen
during the ingestion even before reporting ready-to-serve yet. And right now, the lookup is happening
in consumer thread, and the {@link #adjustStoragePartition} will be invoked in drainer thread.

This code change also added min/max stats to the existing latency metrics in the read path.

TODO:
Evaluate whether we could remove `synchronized` keyword from the functions of `AbstractStorageEngine`.

RB=2680728
BUG=VENG-7921
R=xnma,sdwu,ssen,fvillega,nithakka,anaberez,jlliu,yurzhou,shuhliu,smaji,xyin,xhao,lcli,adietz,zpolicze,mnhuang
A=anaberez,fvillega