[venice-samza] Added API to support logical timestamp for the update from Samza
This code change is trying to leverage the opaque types to keep the current
APIs unchanged, and for the users who need to offer logical timestamp, they
will need to use the new wrapper type: VeniceObjectWithTimestamp.

So far, this code change only support `VeniceSystemProducer#send`, and other
API usages, such as `put` and `delete` are out of scope for now.

This code change did a small GC optimization in VeniceWriter to reuse the
same empty ByteBuffer instance when necessary.

This code change also tweaked the test code for fast-avro to make it
work with JDK11.

RB=2690098
BUG=VENG-7035
R=adietz,xyin,yurzhou,sdwu,shuhliu,ssen,mnhuang,anaberez,jlliu,fvillega,smaji,zpolicze,xhao,xnma,nithakka,lcli
A=ssen,mnhuang