[test] E2E Integration Tests for Active/Active Replication
This RB added 5 end2end integration tests for A/A replication:
1. store with A/A config turned on can consume real-time messages
   from all regions
2. store with A/A config enabled can resolve conflicts within the
   same region based on clock time
3. store with A/A config enabled can resolve conflicts from different
   regions based on clock time
4. store with A/A config enabled can resolve conflicts within the same
   region based on logical timestamp
5. store with A/A config enabled can resolve conflicts from different
   regions based on logical timestamp

All these tests are failing since the related features are not implemented
yet, so a test disable flag is put in each test case; enable the tests once
the feature is enabled. In theory, the above 5 tests will be enabled in order:
firstly, be able to consume from all prod fabrics; secondly, be able to do
proper deterministic-conflict-resolution.

Other changes:
1. Fine tune the logic for hybrid data replication policy safeguard in CreateVersion
   so that store with A/A config enabled can talk to child controllers.
2. Added a new implementation of mock time that returns time based on a list of
   synthetic timestamps

RB=2690392
BUG=VENG-7903
G=voldemort-reviewers
R=ssen,nithakka,goliu,jlliu,smaji,lcli,sdwu,fvillega,shuhliu,xyin,xhao,adietz,yurzhou,xnma,anaberez,zpolicze
A=shuhliu