[da-vinci] Align Da Vinci's version swap policy with classical Venice
This RB aligns Da Vinci's version swap policy with classical Venice as well as being compatible with existing faulty version concept.
With this change, futureVersion will only be swapped to current if and only if:
(1) partitions are ready-to-serve, non-faulty.
(2) version number <= store config's current version's version number.
(3) the version exists in the store config.

The RB only address swapping condition, it does not address bootstarp and rollback. They will be included it the future RB.

RB=2623989
BUG=VENG-7642
G=voldemort-reviewers
R=adietz,goliu,nithakka,ssen,xyin,xnma,zpolicze,yurzhou,sdwu,anaberez,shuhliu,mnhuang,lcli,smaji,xhao,fvillega
A=xyin