[A/A][da-vinci][server] Make StorageService can create TSMD-based partition based on store level A/A config
This RB makes RocksDBStorageEngine aware of the store version level A/A config and use it to determine whether to create TSMD-backed RocksDB storage partition for data partition. StorageService will look up the version config from readOnlyStoreRepository for the A/A config and pass it to RocksDBStorageEngine constructor.

Add new unit test to verify that it can spawn TSMD partition when A/A is enabled.

RB=2698757
G=voldemort-reviewers
R=xnma,xhao,lcli,adietz,mnhuang,fvillega,goliu,yurzhou,sdwu,ssen,xyin,nithakka,smaji,zpolicze,anaberez,shuhliu
A=mnhuang,goliu,ssen