[server][da-vinci] Pass the kafka URL context along with consumer record.
For A/A we will need to identify the kafka colo/URL where a record was
consumed from. Introduced a VeniceConsumerRecord class which extends the
kafka ConsumerRecord. This has the advantage of fetching the kafkaUrl
from the record object itself instead of passing around the kafkaUrl in
many function.
Also added a config which will map the kafka URL to an integer. This is
needed as in KME upstreamKafkaClusterId is an integer field and not
a string.

The scope of this change is to introduce these new fields and config,
these are not used yet.

RB=2704903
BUG=VENG-7967
G=voldemort-reviewers
R=smaji,xnma,nithakka,xhao,xyin,anaberez,goliu,mnhuang,lcli,yurzhou,sdwu,zpolicze,jlliu,adietz,fvillega,shuhliu
A=goliu,mnhuang