[server] Added an IndexedMap
The IndexedMap interface implements Map and several functions for dealing with its
content via the index corresponding to their insertion order. The concept is similar
to the LinkedHashMap, but affords the user a greater control over how the data is
laid out when iterating over it.

The IndexedHashMap is a fork of the standard HashMap which implements the interface.
It is composed of an ArrayList to keep track of the insertion & iteration order of
the entries. The performance of the various operations should be as follows:

- The get and put performance should be constant-time, like HashMap.
- The iteration performance should be equal or slightly better than HashMap, since a
  regular map needs to weed through some empty nodes which are part of the map's
  internal structure.
- The delete performance should be worse than that of the HashMap, since it needs to
  do extra bookkeeping to keep the list and map nodes in sync, with a worse case of O(N).
- The indexOf(key) should be constant-time, like HashMap's get.
- The putByIndex should be O(N) in the worst case.
- The getByIndex should be O(1).
- The removeByIndex should be O(N) in the worst case.
- The moveElement should be O(N) in the worst case.

RB=2710112
BUG=VENG-7998
G=voldemort-reviewers
R=nithakka,zpolicze,ssen,xnma,shuhliu,lcli,adietz,yurzhou,sdwu,xhao,smaji,xyin,anaberez,mnhuang,goliu,jlliu
A=zpolicze