[server] Safefuard for KME/StoreVersionState/PartitionState protocol upgrade in server.
When these protocol evolution happens, controller writes the new
protocol schema in Zk and server reads them. So the expected deployment
order is controller first and then server. But when this order is not
followed due to any reason there will be problem as upgraded servers
will start using the new protocol verison schema which would not be
understood by non-upgraded server and deserialization failures will
happen.
Adding following safeguard in server side to deal with this problem.
When server starts up it will verify the presence of the current
version schema for these protocols. If the current version schema is
not found in ZK then the server will not start up and exit.

RB=2715318
BUG=VENG-7958
G=voldemort-reviewers
R=xnma,mnhuang,jlliu,sdwu,xyin,fvillega,zpolicze,anaberez,yurzhou,adietz,nithakka,smaji,xhao,goliu,lcli,shuhliu
A=mnhuang