[server] Avoid doing schema presence check in dev deployment
Since in dev deployment the venice router may not be running
schemaReader will not be able to fetch the schema. This will make
server not start up correctly. Also we don't need to validate this
schema presence scheck in dev environment. So introducing a config
to skip this check completely.

RB=2722292
BUG=VENG-7958
G=voldemort-reviewers
R=adietz,xhao,lcli,xnma,smaji,anaberez,shuhliu,xyin,fvillega,sdwu,nithakka,goliu,mnhuang,jlliu,zpolicze,yurzhou
A=goliu