[test-only][controller] Fix flaky HAASController tests
This RB fixes both `testTransitionToHAASControllerAsControllerClusterLeader` and `testTransitionToHAASControllerAsStorageClusterLeader` in `TestHAASController` class. When controller cluster is performing rolling bounce, newly added controller might not be fully initilized -- Spark server is able t o receive END_OF_PUSH command but Venice Helix Resource is not ready and it will throw VeniceNoClusterException as the ControllerResponse.

The fix is simple: We add a `TestUtils.waitForNonDeterministicAssertion` to wrap the EOP request to retry until success.

RB=2732524
BUG=VENG-7362
G=voldemort-reviewers
R=lcli,xyin,anaberez,akozma,mnhuang,shuhliu,goliu,xhao,xnma,yurzhou,ssen,adietz,sdwu,smaji,zpolicze,nithakka,fvillega
A=lcli