[controller] Support KME protocol forward compatibility in controller
Currently controller KafkaValueSerializer does not have the schemaReader
set. This does not allow the controller to read a new KME schema
version dynamically. If there is a new KME version deployed and due to any reason the
controller needs to be rolled back then it would not be able to
deserialize a message which was written using a later KME version.

This change wires in the schemaReader to the kafka value serializer that
controller uses. This will help controller download the new KME schema
to deserialize a message written using later schema version.

RB=2729587
BUG=VENG-7980
G=voldemort-reviewers
R=yurzhou,sdwu,xyin,shuhliu,xhao,fvillega,anaberez,goliu,mnhuang,zpolicze,lcli,nithakka,xnma,jlliu,smaji,adietz
A=mnhuang