[da-vinci][test-common][common] Rewrite ingestion isolation initialization logics
This RB follows the decision made in ingesiton isolation security enhancement to rewrite the initialization logics.
Instead of forking the process and then send the INIT command to forked process with all related configs, we now saves all configs to a temp config file and ask the forked process to load configs from it during start() call. With this approach, server will be able to detect whether it is SSL enabledduring start up and spawn with correct Netty server.

Note that this diff only contains the logic to save and load configs, it does not contain SSL logics for Netty server. SSL enhancement will be in second RB.

Also, this RB adjusts some tests to reflect the changes.

RB=2738852